add color and death packet
